+++ lib/core/widgets/app_router.dart
@@ -5,6 +5,7 @@ import 'package:browny_applications_new/core/widgets/app_text.dart';
 import 'package:browny_applications_new/feature/articles/screens/article_detail_page.dart';
 import 'package:browny_applications_new/feature/browny_shop/screens/browny_shop_page.dart';
 import 'package:browny_applications_new/feature/browny_shop/screens/browny_shop_product_detail_page.dart';
+import 'package:browny_applications_new/feature/browny_shop/screens/browny_shop_search_page.dart';
 import 'package:browny_applications_new/feature/browny_shop/screens/browny_shop_cart_page.dart';
 import 'package:browny_applications_new/feature/browny_shop/screens/browny_shop_selected_page.dart';
 import 'package:browny_applications_new/feature/browny_shop/screens/receipt_browny_shop_page.dart';
@@ -98,6 +99,11 @@ class AppRouter {
               state.pathParameters[BrownyShopProductDetailPage.kProductId] ??
               '',
         ),
+      ),
+      GoRoute(
+        path: BrownyShopSearchPage.pagePath,
+        name: BrownyShopSearchPage.pageName,
+        builder: (context, state) => const BrownyShopSearchPage(),
       ),
       GoRoute(
         path: BrownyShopCartPage.pagePath,